requirements.txt
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
conda==4.3.16
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
importlib-resources==5.1.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2020.8.13
platformdirs==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==4.0.1
pycosat==0.6.3
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
six==1.16.0
smmap==4.0.0
streamlit==0.84.0
streamlit-metrics==0.1.0
sweetviz==2.1.2
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.25.10
validators==0.18.2
virtualenv==20.0.31
virtualenv-clone==0.5.4
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
XlsxWriter==1.4.4
zipp==3.4.1
